Move games.js to the commands folder
